Move code from Parser/Crayon/Crayon*Parser.cs to Parser/Abstract*Parser.cs
