+++ Compiler/Parser/Acrylic/AcrylicExpressionParser.cs
@@ -4,11 +4,9 @@ namespace Parser.Acrylic
 {
     internal class AcrylicExpressionParser : AbstractExpressionParser
     {
-        private ParserContext parser;
         public AcrylicExpressionParser(ParserContext parser)
-        {
-            this.parser = parser;
-        }
+            : base(parser)
+        { }
 
         internal override Expression Parse(TokenStream tokens, Node owner)
         {
