+++ Compiler/Parser/Crayon/CrayonAnnotationParser.cs
@@ -1,56 +1,8 @@
-﻿using Parser.ParseTree;
-using System.Collections.Generic;
-
-namespace Parser.Crayon
+﻿namespace Parser.Crayon
 {
     internal class CrayonAnnotationParser : AbstractAnnotationParser
     {
-        private ParserContext parser;
         public CrayonAnnotationParser(ParserContext parser)
-        {
-            this.parser = parser;
-        }
-
-        internal override AnnotationCollection ParseAnnotations(TokenStream tokens)
-        {
-            AnnotationCollection annotationCollection = new AnnotationCollection(this.parser);
-            while (tokens.IsNext("@"))
-            {
-                annotationCollection.Add(this.ParseAnnotation(tokens));
-            }
-
-            annotationCollection.Validate();
-
-            return annotationCollection;
-        }
-
-        private Annotation ParseAnnotation(TokenStream tokens)
-        {
-            Token annotationToken = tokens.PopExpected("@");
-            Token typeToken = tokens.Pop();
-
-            // TODO: refactor this. All built-in annotations should be exempt from the VerifyIdentifier check in an extensible way.
-            if (typeToken.Value != this.parser.Keywords.PRIVATE)
-            {
-                parser.VerifyIdentifier(typeToken);
-            }
-
-            Annotation annotation = new Annotation(annotationToken, typeToken);
-            List<Expression> args = new List<Expression>();
-            if (tokens.PopIfPresent("("))
-            {
-                while (!tokens.PopIfPresent(")"))
-                {
-                    if (args.Count > 0)
-                    {
-                        tokens.PopExpected(",");
-                    }
-
-                    args.Add(this.parser.ExpressionParser.Parse(tokens, annotation));
-                }
-            }
-            annotation.SetArgs(args);
-            return annotation;
-        }
+            : base(parser) { }
     }
 }
