+++ Compiler/Pastel/Transpilers/CTranslator.cs
@@ -1,6 +1,5 @@
 ﻿using Pastel.Nodes;
 using System;
-using System.Collections.Generic;
 
 namespace Pastel.Transpilers
 {
